Renamed app packag and app name
