+++ android/app/src/main/AndroidManifest.xml
@@ -1,18 +1,10 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.scifi">
+  package="com.movie_intent">
 
     <uses-permission android:name="android.permission.INTERNET" />
-      <supports-screens 
-        android:resizeable="true"
-        android:smallScreens="true"
-        android:normalScreens="true"
-        android:largeScreens="true"
-        android:anyDensity="true"
-        />
+
     <application
-     android:largeHeap="true"
       android:name=".MainApplication"
-      android:hardwareAccelerated="true"
       android:label="@string/app_name"
       android:icon="@mipmap/ic_launcher"
       android:roundIcon="@mipmap/ic_launcher_round"
