+++ android/app/src/main/java/com/movie_intent/MainActivity.java
@@ -1,4 +1,4 @@
-package com.scifi;
+package com.movie_intent;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "scifi";
+    return "Movie Intent";
   }
 }
