+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">scifi</string>
+    <string name="app_name">Movie Intent</string>
 </resources>
