+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'scifi'
+rootProject.name = 'Movie Intent'
 include ':lottie-react-native'
 project(':lottie-react-native').projectDir = new File(rootProject.projectDir, '../node_modules/lottie-react-native/src/android')
 include ':react-native-vector-icons'
