+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "scifi",
-  "displayName": "scifi"
+  "name": "Movie Intent",
+  "displayName": "Movie Intent"
 }
