+++ ios/Movie Intent.xcodeproj/xcshareddata/xcschemes/Movie Intent-tvOS.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "scifi-tvOS.app"
-               BlueprintName = "scifi-tvOS"
-               ReferencedContainer = "container:scifi.xcodeproj">
+               BuildableName = "Movie Intent-tvOS.app"
+               BlueprintName = "Movie Intent-tvOS"
+               ReferencedContainer = "container:Movie Intent.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "scifi-tvOSTests.xctest"
-               BlueprintName = "scifi-tvOSTests"
-               ReferencedContainer = "container:scifi.xcodeproj">
+               BuildableName = "Movie Intent-tvOSTests.xctest"
+               BlueprintName = "Movie Intent-tvOSTests"
+               ReferencedContainer = "container:Movie Intent.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "scifi-tvOS.app"
-            BlueprintName = "scifi-tvOS"
-            ReferencedContainer = "container:scifi.xcodeproj">
+            BuildableName = "Movie Intent-tvOS.app"
+            BlueprintName = "Movie Intent-tvOS"
+            ReferencedContainer = "container:Movie Intent.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "scifi-tvOS.app"
-            BlueprintName = "scifi-tvOS"
-            ReferencedContainer = "container:scifi.xcodeproj">
+            BuildableName = "Movie Intent-tvOS.app"
+            BlueprintName = "Movie Intent-tvOS"
+            ReferencedContainer = "container:Movie Intent.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
