+++ ios/Movie Intent.xcodeproj/xcshareddata/xcschemes/Movie Intent.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "scifi.app"
-               BlueprintName = "scifi"
-               ReferencedContainer = "container:scifi.xcodeproj">
+               BuildableName = "Movie Intent.app"
+               BlueprintName = "Movie Intent"
+               ReferencedContainer = "container:Movie Intent.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "scifiTests.xctest"
-               BlueprintName = "scifiTests"
-               ReferencedContainer = "container:scifi.xcodeproj">
+               BuildableName = "Movie IntentTests.xctest"
+               BlueprintName = "Movie IntentTests"
+               ReferencedContainer = "container:Movie Intent.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "scifi.app"
-            BlueprintName = "scifi"
-            ReferencedContainer = "container:scifi.xcodeproj">
+            BuildableName = "Movie Intent.app"
+            BlueprintName = "Movie Intent"
+            ReferencedContainer = "container:Movie Intent.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "scifi.app"
-            BlueprintName = "scifi"
-            ReferencedContainer = "container:scifi.xcodeproj">
+            BuildableName = "Movie Intent.app"
+            BlueprintName = "Movie Intent"
+            ReferencedContainer = "container:Movie Intent.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
