+++ ios/Movie IntentTests/Movie IntentTests.m
@@ -7,11 +7,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React"
 
-@interface scifiTests : XCTestCase
+@interface Movie IntentTests : XCTestCase
 
 @end
 
-@implementation scifiTests
+@implementation Movie IntentTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
