+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "scifi",
+  "name": "Movie_Intent",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -21,6 +21,7 @@
     "react-native": "0.63",
     "react-native-appearance": "^0.3.4",
     "react-native-dotenv": "^2.4.0",
+    "react-native-eject": "^0.1.2",
     "react-native-elements": "^2.3.1",
     "react-native-gesture-handler": "^1.8.0",
     "react-native-reanimated": "^1.13.0",
